Implement note computation
InstrumentModel->getStringNote now works.

Additionally, I changed the call to return a regular char instead
of a char pointer for the sake of simplicity.  I will probably
change this back later on.

Changed the frequency to a double instead of a regular float.
This let me use the fmod operation in tgmath/cmath for convenient
mod operations.