Begin switch to boost
Decided to handle json parsing with libboost, as I would rather not
spend a lot of time building a finite state machine just to
reinvent the wheel.

This commit basically makes changes to the stub of a json reading
method to make it read in and begin parsing with boost.

TODO: make better use of the boost parsing (go deeper into the
tree)